requirements.txt
python-dotenv~=1.0.0
langchain~=0.0.187
chromadb~=0.3.25
sentence-transformers
openai